add kafka integration

+++ adapter/api/src/main/java/uk/co/mruoc/nac/api/dto/ApiBoard.java
@@ -12,6 +12,6 @@ import lombok.RequiredArgsConstructor;
 @Data
 public class ApiBoard {
 
-  private final long size;
+  private final int size;
   private final Collection<ApiLocation> locations;
 }
